// src/api/generated/models/activityView.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */
import type { ActivityViewerAction } from './activityViewerAction';

export interface ActivityView {
  available_actions: ActivityViewerAction[];
  body: string;
  capacity: number;
  contact: string;
  contact_type: string;
  created_at: string;
  created_by: number;
  end_at: string;
  id: number;
  location: string;
  registered_count: number;
  /** @nullable */
  review_comment?: string | null;
  review_status: string;
  signup_end_at: string;
  signup_start_at: string;
  start_at: string;
  status: string;
  summary: string;
  title: string;
  updated_at: string;
  updated_by: number;
  version: number;
  viewer_relation: string;
}