// src/api/generated/models/tokenPair.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Campus Platform API
 * OpenAPI spec version: 1.1.0
 */
import type { TokenPairTokenType } from './tokenPairTokenType';

export interface TokenPair {
  access_token: string;
  refresh_token: string;
  token_type: TokenPairTokenType;
  /** @minimum 1 */
  expires_in: number;
}